call eldarica by shell in extraction and prediction

+++ Heuristic_selection/src/utils.py
@@ -40,38 +40,49 @@ def plot_scatter(true_Y,predicted_Y,name="",range=0):
     plt.clf()
 
 
-def run_eldarica_with_shell_pool(filePath, fun, eldarica_parameters, countinous_extract=True):
+def run_eldarica_with_shell_pool(filePath, fun, eldarica_parameters,timeout=60,thread=4,countinous_extract=True):
     file_list = []
     for root, subdirs, files in os.walk(filePath):
         if len(subdirs) == 0:
             if countinous_extract == True:
                 for file in glob.glob(root + "/*.smt2"):
                     if not os.path.exists(file + ".circles.gv"):
-                        file_list.append([file, eldarica_parameters])
+                        file_list.append([file,eldarica_parameters,timeout])
             else:
                 for file in glob.glob(root + "/*.smt2"):
-                    file_list.append([file, eldarica_parameters])
-    pool = Pool(processes=4)
+                    file_list.append([file,eldarica_parameters,timeout])
+    run_eldarica_with_shell_pool_with_file_list(thread,fun,file_list)
+
+def run_eldarica_with_shell_pool_with_file_list(thread,fun,file_list):
+    pool = Pool(processes=thread)
     pool.map(fun, file_list)
 
 
 def run_eldarica_with_shell(file_and_param):
-    file = file_and_param[0]
-    eldarica = "../eldarica-graph-generation/eld "
-    # file = "../benchmarks/ulimit-test/Problem19_label06_true-unreach-call.c.flat_000.smt2"
-    file_name = file[file.rfind("/") + 1:]
-    parameter_list = file_and_param[1]
-    shell_file_name = "run-ulimit" + "-" + file_name + ".sh"
-    timeout_command = "timeout 30"
-    f = open(shell_file_name, "w")
-    f.write("#!/bin/sh\n")
-    # f.write("ulimit -m 4000000; \n")
-    # f.write("ulimit -v 6000000; \n")
-    # f.write("ulimit -a; \n")
-    f.write(timeout_command + " " + eldarica + " " + file + " " + parameter_list + "\n")
-    f.close()
-    supplementary_command = ["sh", shell_file_name]
-    eld = subprocess.Popen(supplementary_command, stdout=subprocess.DEVNULL, shell=False)
-    eld.wait()
-    # subprocess.call(supplementary_command)
-    os.remove(shell_file_name)
+    if len(file_and_param)!=0:
+        file = file_and_param[0]
+        eldarica = "../eldarica-graph-generation/eld "
+        # file = "../benchmarks/ulimit-test/Problem19_label06_true-unreach-call.c.flat_000.smt2"
+        file_name = file[file.rfind("/") + 1:]
+        parameter_list = file_and_param[1]
+        timeout=file_and_param[2]
+        shell_file_name = "run-ulimit" + "-" + file_name + ".sh"
+        timeout_command = "timeout "+str(timeout)
+        f = open(shell_file_name, "w")
+        f.write("#!/bin/sh\n")
+        # f.write("ulimit -m 4000000; \n")
+        # f.write("ulimit -v 6000000; \n")
+        # f.write("ulimit -a; \n")
+        f.write(timeout_command + " " + eldarica + " " + file + " " + parameter_list + "\n")
+        f.close()
+        supplementary_command = ["sh", shell_file_name]
+        print("extracting "+file_name)
+        eld = subprocess.Popen(supplementary_command, stdout=subprocess.DEVNULL, shell=False)
+        eld.wait()
+        print("extracting " + file_name + " finished")
+        # subprocess.call(supplementary_command)
+        os.remove(shell_file_name)
+
+        if not os.path.exists(file + ".circles.gv"):
+            os.rename(file,"../benchmarks/shell-timemout/"+file_name)
+            print("extracting " + file_name + " failed due to time out, move file to shell-timemout")
